Fix tests and notebooks and everything
